+++ service/scheduler/dao/vote_dao.go
@@ -14,6 +14,9 @@ import (
 // InsertVoteCloud insert cloud into vote cloud collection
 func InsertVoteCloud(col *mongo.Collection, cloud entity.VoteCloud) error {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
+	if colErr := VerifyCollection(col); colErr != nil {
+		return colErr
+	}
 	_, err := col.InsertOne(context.TODO(), cloud)
 	if err != nil {
 		return err
@@ -24,6 +27,9 @@ func InsertVoteCloud(col *mongo.Collection, cloud entity.VoteCloud) error {
 // CloudsCount get the number of clouds whose id is cid
 func CloudsCount(col *mongo.Collection, cid string) (int64, error) {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
+	if colErr := VerifyCollection(col); colErr != nil {
+		return -1, colErr
+	}
 	count, err := col.CountDocuments(context.TODO(), bson.M{"cloud_id": cid})
 	if err != nil {
 		return count, err
@@ -35,7 +41,9 @@ func CloudsCount(col *mongo.Collection, cid string) (int64, error) {
 // DeleteVoteCloud delete the cloud
 func DeleteVoteCloud(col *mongo.Collection, id string) error {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
-
+	if colErr := VerifyCollection(col); colErr != nil {
+		return colErr
+	}
 	_, err := col.DeleteOne(context.TODO(), bson.M{"cloud_id": id})
 	if err != nil {
 		return err
@@ -47,7 +55,9 @@ func DeleteVoteCloud(col *mongo.Collection, id string) error {
 // AddVoteNum add vote number
 func AddVoteNum(col *mongo.Collection, vote int, id string) (int, error) {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
-
+	if colErr := VerifyCollection(col); colErr != nil {
+		return -1, colErr
+	}
 	res, err := col.UpdateOne(
 		context.TODO(),
 		bson.M{"cloud_id": id},
@@ -68,6 +78,9 @@ func GetVoteCloud(col *mongo.Collection, id string) (entity.VoteCloud, error) {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
 
 	var result entity.VoteCloud
+	if colErr := VerifyCollection(col); colErr != nil {
+		return result, colErr
+	}
 	err := col.FindOne(context.TODO(), bson.M{"cloud_id": id}).Decode(&result)
 	if err != nil {
 		return result, err
@@ -81,6 +94,9 @@ func GetAllVoteCloud(col *mongo.Collection) ([]entity.VoteCloud, error) {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
 
 	var result []entity.VoteCloud
+	if colErr := VerifyCollection(col); colErr != nil {
+		return result, colErr
+	}
 	cur, err := col.Find(context.TODO(), bson.M{})
 	defer cur.Close(context.TODO())
 	if err != nil {
@@ -100,6 +116,9 @@ func GetAllVoteCloud(col *mongo.Collection) ([]entity.VoteCloud, error) {
 // GetVoteNumber : Get the vote number of the cloud with id
 func GetVoteNumber(col *mongo.Collection, id string) (int, error) {
 	//col := d.client.Database(d.database).Collection(d.cloudCollection)
+	if colErr := VerifyCollection(col); colErr != nil {
+		return -1, colErr
+	}
 	var result entity.VoteCloud
 	err := col.FindOne(context.TODO(), bson.M{"cloud_id": id}).Decode(&result)
 	if err != nil {
